fix: Resolve cargo build errors for release binary
- Fix wasmtime API compatibility in wasm_runtime.rs (WasiView trait, ResourceLimiter signature)
- Convert SQLx compile-time queries to runtime queries to avoid DATABASE_URL requirement
- Fix type mismatches (i32 vs i64 for artifact_count in backup_service)
- Add WasmError variant to AppError enum with proper From implementation
- Add activate_plugin_at_startup public method for main.rs plugin loading
- Fix migration empty array type casting (ARRAY[]::TEXT[])
- Rename conflicting migration files (014 -> 016, 015 -> 017)
- Add base64 and move tempfile to regular dependencies
- Add GenericHandler::new() method and Default impl

+++ backend/src/api/handlers/admin.rs
@@ -58,7 +58,7 @@ pub struct BackupResponse {
     pub status: String,
     pub storage_path: Option<String>,
     pub size_bytes: i64,
-    pub artifact_count: i32,
+    pub artifact_count: i64,
     pub started_at: Option<chrono::DateTime<chrono::Utc>>,
     pub completed_at: Option<chrono::DateTime<chrono::Utc>>,
     pub error_message: Option<String>,
